Backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.2.2
django-environ>=0.11
django-cors-headers>=4.4
psycopg2-binary>=2.9
gunicorn>=22.0